fix(kb): support fractional crafting_time in items generator
The 2026-06-19 game data made crafting_time a real duration in seconds
(was a dead field; crafting was always 1 tick/10s). The crafting DB now
stores fractional values, so scanning crafting_time into a Go int crashed
the generator. Switch CraftingTime to float64 across the recipe/produced-by/
built-by/facility structs and the scan, and relabel recipe-page "ticks" as
"s" since the values are now real seconds (2145 ticks would misread as ~6h).

+++ cmd/generate-items-kb/facilities.go
@@ -76,7 +76,7 @@ type RecipeSummary struct {
 	ID           string        `json:"id"`
 	Name         string        `json:"name"`
 	Category     string        `json:"category"` // Required for linking to recipe pages
-	CraftingTime int           `json:"crafting_time"`
+	CraftingTime float64       `json:"crafting_time"`
 	Inputs       []MaterialRef `json:"inputs"`
 	Outputs      []MaterialRef `json:"outputs"`
 }
@@ -632,7 +632,7 @@ func validateFacilityRecipes(facilities map[string]*Facility, recipes map[string
 
 		// Compare crafting time
 		if fac.Recipe.CraftingTime != recipe.CraftingTime {
-			log.Printf("warning: facility %s recipe %s crafting time mismatch: facility has %d, recipe has %d",
+			log.Printf("warning: facility %s recipe %s crafting time mismatch: facility has %g, recipe has %g",
 				fac.ID, fac.Recipe.ID, fac.Recipe.CraftingTime, recipe.CraftingTime)
 		}
 	}
